backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy>=1.26
opencv-python
pydantic
python-dotenv==1.0.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
openai==1.54.0
google-search-results==2.4.2
# PyTorch CPU-optimized versions
# For CPU-only systems, install with: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# torch>=2.0.0
# torchvision>=0.15.0
# timm>=0.9.0
# psutil>=5.9.0